edit  instractor view progress page
47 Instructor Analytics: Reports on content effectiveness:
a. Rating for each module in the course.
b. Rating for the course in general. (Average of modules' ratings)

46 Instructor Analytics:Reports on student engagement:a.Number of enrolled students in a course.
b. Number of students completed the course.
c. Numbers of students based on their performance metric (Below average, Average,
Above Average, Excellent).

and added a button  complete module  that create progress and added modules in it